requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
streamlit>=1.35.0
python-dotenv>=1.0.0

# ── LLM: Groq (FREE - get key at console.groq.com, no credit card) ───────────
groq>=0.9.0

# ── Embeddings: runs 100% locally, no API key needed ─────────────────────────
sentence-transformers>=3.0.0

# ── RAG Vector Store ──────────────────────────────────────────────────────────
faiss-cpu>=1.7.4
numpy>=1.26.0

# ── PDF parsing ───────────────────────────────────────────────────────────────
pdfplumber>=0.11.0

# ── Web Search (free DuckDuckGo by default) ───────────────────────────────────
duckduckgo-search>=6.1.0
requests>=2.31.0
